End of day Sunday
Display for launchpad buttons is working, at least to the point where a button is pressed and the display lights up (doesnt turn off yet).  Design seems to be relatively error free, except for some timing constraint shenanigans and something in song_reader.  Need to investigate timing constraints to see whether anything needs to be fixed, or if its our starter code.